Agentic_workflow/stock-agent-service/requirements.txt
langgraph
langchain
langchain-google-genai
yfinance
tavily-python
python-dotenv
pytest
